requirements.txt
requests
python-dotenv
google-cloud-storage
google-auth